+++ test/providers/openai/index.test.ts
@@ -1,17 +1,15 @@
 import OpenAI from 'openai';
-import { disableCache, enableCache, fetchWithCache } from '../../src/cache';
+import { disableCache, enableCache, fetchWithCache } from '../../../src/cache';
 import {
   calculateOpenAICost,
-  failApiCall,
-  getTokenUsage,
   OpenAiChatCompletionProvider,
   OpenAiCompletionProvider,
   OpenAiEmbeddingProvider,
   OpenAiGenericProvider,
-} from '../../src/providers/openai';
-import { OpenAiAssistantProvider } from '../../src/providers/openai/assistant';
+} from '../../../src/providers/openai';
+import { OpenAiAssistantProvider } from '../../../src/providers/openai/assistant';
 
-jest.mock('../../src/cache');
+jest.mock('../../../src/cache');
 
 const mockFetchWithCache = jest.mocked(fetchWithCache);
 
@@ -25,79 +23,6 @@ describe('OpenAI Provider', () => {
 
   afterEach(() => {
     enableCache();
-  });
-
-  describe('failApiCall', () => {
-    it('should format OpenAI API errors', () => {
-      const error = new OpenAI.APIError(400, {}, 'Bad request', {});
-      Object.defineProperty(error, 'type', {
-        value: 'invalid_request_error',
-        writable: true,
-        configurable: true,
-      });
-      Object.defineProperty(error, 'message', {
-        value: 'Bad request',
-        writable: true,
-        configurable: true,
-      });
-      Object.defineProperty(error, 'status', {
-        value: 400,
-        writable: true,
-        configurable: true,
-      });
-
-      const result = failApiCall(error);
-      expect(result).toEqual({
-        error: `API error: invalid_request_error 400 Bad request`,
-      });
-    });
-
-    it('should format generic errors', () => {
-      const error = new Error('Network error');
-      const result = failApiCall(error);
-      expect(result).toEqual({
-        error: `API error: Error: Network error`,
-      });
-    });
-  });
-
-  describe('getTokenUsage', () => {
-    it('should return token usage for non-cached response', () => {
-      const data = {
-        usage: {
-          total_tokens: 100,
-          prompt_tokens: 40,
-          completion_tokens: 60,
-        },
-      };
-
-      const result = getTokenUsage(data, false);
-      expect(result).toEqual({
-        total: 100,
-        prompt: 40,
-        completion: 60,
-      });
-    });
-
-    it('should return cached token usage', () => {
-      const data = {
-        usage: {
-          total_tokens: 100,
-        },
-      };
-
-      const result = getTokenUsage(data, true);
-      expect(result).toEqual({
-        cached: 100,
-        total: 100,
-      });
-    });
-
-    it('should handle missing usage data', () => {
-      const data = {};
-      const result = getTokenUsage(data, false);
-      expect(result).toEqual({});
-    });
   });
 
   describe('OpenAiGenericProvider', () => {
